refactor(CodeBlock): rename concrete code blocks, refactor the hierarchy
